feat(user-profile): implement PublicUserProfile component for user profile viewing
